+++ pyro/distributions/transforms/discrete_cosine.py
@@ -73,3 +73,8 @@ class DiscreteCosineTransform(Transform):
 
     def log_abs_det_jacobian(self, x, y):
         return x.new_zeros((1,) * self.event_dim)
+
+    def with_cache(self, cache_size=1):
+        if self._cache_size == cache_size:
+            return self
+        return DiscreteCosineTransform(-self.event_dim, self.smooth, cache_size)
